micro-optimizations
