+++ .Rbuildignore
@@ -10,3 +10,4 @@
 ^CRAN-SUBMISSION$
 ^doc$
 ^Meta$
+^R/utils\.R$
