+++ NAMESPACE
@@ -12,7 +12,6 @@ export(inject_na)
 export(knn_imp)
 export(mean_imp_col)
 export(pca_imp)
-export(qlogis_clamped)
 export(sim_mat)
 export(slide_imp)
 export(tune_imp)
