--- man/qlogis_clamped.Rd
@@ -1,31 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utils.R
-\name{qlogis_clamped}
-\alias{qlogis_clamped}
-\title{Clamped Quantile Function of the Logistic Distribution}
-\usage{
-qlogis_clamped(obj, eps = 1e-10, ...)
-}
-\arguments{
-\item{obj}{A numeric matrix with values in between 0 and 1.}
-
-\item{eps}{A single positive number giving the clamping margin. Input values
-are squeezed into the interval (eps, 1 - eps) before \code{\link[stats:Logistic]{stats::qlogis()}}. Defaults
-to \code{1e-10}.}
-
-\item{...}{Additional arguments passed to \code{\link[stats:Logistic]{stats::qlogis()}}.}
-}
-\value{
-A numeric matrix of the same dimensions as \code{obj} containing
-logit-transformed values. The output is guaranteed to be finite.
-}
-\description{
-A wrapper around \code{\link[stats:Logistic]{stats::qlogis()}} that clamps input probabilities
-away from 0 and 1 by \code{eps} to avoid infinite return values. Useful for converting
-beta values into M-values.
-}
-\examples{
-p <- matrix(c(0, 0.01, 0.5, 0.99, 1, NA), nrow = 2)
-qlogis_clamped(p)
-
-}
