+++ man/sim_mat.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utils.R
+% Please edit documentation in R/sim_mat.R
 \name{sim_mat}
 \alias{sim_mat}
 \title{Simulate Methylation Beta Values with Metadata}
