+++ src/fastSVD.cpp
@@ -1,185 +1,5 @@
-#include "partial_eig.h"
-// #include <rcpptimer.h>
-
-constexpr double PCA_TOL = 1e-15;
-
-// helpers to address the heavy overhead in svd copy creating functions like %=
-static inline void scale_cols_inplace(double *M, const double *s,
-                                      arma::uword nr, arma::uword nc)
-{
-  for (arma::uword j = 0; j < nc; ++j)
-  {
-    const double sj = s[j];
-    double *col = M + j * nr;
-    for (arma::uword i = 0; i < nr; ++i)
-    {
-      col[i] *= sj;
-    }
-  }
-}
-
-static inline void scale_rows_inplace(double *M, const double *w,
-                                      arma::uword nr, arma::uword nc)
-{
-  for (arma::uword j = 0; j < nc; ++j)
-  {
-    double *col = M + j * nr;
-    for (arma::uword i = 0; i < nr; ++i)
-    {
-      col[i] *= w[i];
-    }
-  }
-}
-
-// Copy source -> dest while scaling row i by w(i): dst(i,j) = src(i,j) * w(i).
-static inline void copy_scale_rows(double *dst, const double *src,
-                                   const double *w,
-                                   arma::uword nr, arma::uword nc)
-{
-  for (arma::uword j = 0; j < nc; ++j)
-  {
-    double *dcol = dst + j * nr;
-    const double *scol = src + j * nr;
-    for (arma::uword i = 0; i < nr; ++i)
-    {
-      dcol[i] = scol[i] * w[i];
-    }
-  }
-}
-
-// ---------------------------------------------------------------------------
-// SVD with symmetric eigs_sym on gram matrix
-// ---------------------------------------------------------------------------
-inline void SVD_triplet(const arma::mat &Xhat,
-                        const double *sw,
-                        const double *isw,
-                        const arma::uword ncp,
-                        const bool tall,
-                        arma::vec &vs_top,
-                        double &trace_val,
-                        arma::mat &U,
-                        arma::mat &V,
-                        arma::mat &X_work,
-                        arma::mat &AA_NxN,
-                        arma::vec &d_inv,
-                        arma::mat &Vn,
-                        arma::vec &eigvals,
-                        arma::mat &eigvecs,
-                        GramWorkspace &gram_ws)
-{
-  // Rcpp::Timer timer("SVD_triplet");
-  const arma::uword nr = static_cast<arma::uword>(gram_ws.nr);
-  const arma::uword nc = static_cast<arma::uword>(gram_ws.nc);
-
-  // timer.tic("form_gram");
-  if (tall)
-  {
-    // X_work = diag(sqrt_w) * Xhat, fused copy+scale (single pass)
-    X_work.set_size(nr, nc);
-    copy_scale_rows(X_work.memptr(), Xhat.memptr(), sw, nr, nc);
-    syrk_upper(X_work, AA_NxN, gram_ws); // AA = X_work^T X_work
-  }
-  else
-  {
-    syrk_upper(Xhat, AA_NxN, gram_ws); // AA = Xhat Xhat^T (upper)
-    // Apply D * AA * D on upper triangle only.
-    double *M = AA_NxN.memptr();
-    const arma::uword ng = AA_NxN.n_rows;
-    for (arma::uword j = 0; j < ng; ++j)
-    {
-      const double sj = sw[j];
-      double *col = M + j * ng;
-      for (arma::uword i = 0; i <= j; ++i)
-      {
-        col[i] *= sw[i] * sj;
-      }
-    }
-  }
-  // timer.toc("form_gram");
-
-  trace_val = arma::trace(AA_NxN);
-
-  // timer.tic("eig");
-  if (!partial_eig_sym(eigvals, eigvecs, AA_NxN, gram_ws))
-  {
-    Rcpp::stop("`partial_eig_sym` failed to converge");
-  }
-  // timer.toc("eig");
-
-  // Fused clamp+sqrt into vs_top, and build d_inv in the same pass.
-  const arma::uword ke = eigvals.n_elem; // = gram_ws.topk
-  vs_top.set_size(ke);
-  d_inv.set_size(ncp);
-  {
-    const double *ep = eigvals.memptr();
-    double *vp = vs_top.memptr();
-    double *dp = d_inv.memptr();
-    for (arma::uword i = 0; i < ke; ++i)
-    {
-      const double e = ep[i];
-      vp[i] = (e > 0.0) ? std::sqrt(e) : 0.0;
-    }
-    for (arma::uword i = 0; i < ncp; ++i)
-    {
-      dp[i] = (vp[i] > PCA_TOL) ? (1.0 / vp[i]) : 0.0;
-    }
-  }
-
-  // timer.tic("recover_factors");
-  if (tall)
-  {
-    // V = eigvecs[:,0:ncp]; Vn(i,j) = V(i,j) * d_inv(j).
-    const arma::uword n = eigvecs.n_rows;
-    V.set_size(n, ncp);
-    Vn.set_size(n, ncp);
-    const double *EV = eigvecs.memptr();
-    double *Vp = V.memptr();
-    double *Vnp = Vn.memptr();
-    const double *dp = d_inv.memptr();
-    for (arma::uword j = 0; j < ncp; ++j)
-    {
-      const double dj = dp[j];
-      const double *ecol = EV + j * n;
-      double *vcol = Vp + j * n;
-      double *ncol = Vnp + j * n;
-      for (arma::uword i = 0; i < n; ++i)
-      {
-        const double e = ecol[i];
-        vcol[i] = e;
-        ncol[i] = e * dj;
-      }
-    }
-    U = X_work * Vn;
-  }
-  else
-  {
-    // U = eigvecs[:,0:ncp]; Vn(i,j) = U(i,j) * d_inv(j) * sqrt_w(i). Fused.
-    const arma::uword n = eigvecs.n_rows; // == nr
-    U.set_size(n, ncp);
-    Vn.set_size(n, ncp);
-    const double *EV = eigvecs.memptr();
-    double *Up = U.memptr();
-    double *Vnp = Vn.memptr();
-    const double *dp = d_inv.memptr();
-    for (arma::uword j = 0; j < ncp; ++j)
-    {
-      const double dj = dp[j];
-      const double *ecol = EV + j * n;
-      double *ucol = Up + j * n;
-      double *ncol = Vnp + j * n;
-      for (arma::uword i = 0; i < n; ++i)
-      {
-        const double e = ecol[i];
-        ucol[i] = e;
-        ncol[i] = e * dj * sw[i];
-      }
-    }
-    V = Xhat.t() * Vn;
-  }
-  // Undo row weighting on U in place.
-  scale_rows_inplace(U.memptr(), isw, U.n_rows, U.n_cols);
-  // timer.toc("recover_factors");
-}
+#include "eig_sym_sel.h"
+#include "loc_timer.h"
 
 // ---------------------------------------------------------------------------
 // eliminate the scale branch so inner loops auto-vectorize in the while hot loop
@@ -190,49 +10,57 @@ void restandardize(arma::mat &Xhat,
                    arma::rowvec &et,
                    const double *wptr,
                    const arma::uword nrX,
-                   const arma::uword ncX)
+                   const arma::uvec &miss_cols,
+                   const arma::uvec &miss_rows_flat,
+                   const arma::uvec &miss_rows_offsets,
+                   const double *obs_sum,
+                   const double *obs_sumsq)
 {
   double *xptr = Xhat.memptr();
-  for (arma::uword j = 0; j < ncX; ++j)
+  const arma::uword n_mc = miss_cols.n_elem;
+  for (arma::uword cidx = 0; cidx < n_mc; ++cidx)
   {
+    const arma::uword j = miss_cols[cidx];
     double *col = xptr + j * nrX;
     const double mj = mean_p(j);
-    double ej;
+    const double ej = Scale ? et(j) : 1.0;
+    // pass 1: unstandardize entire column
     if constexpr (Scale)
     {
-      ej = et(j);
+      for (arma::uword i = 0; i < nrX; ++i)
+      {
+        col[i] = col[i] * ej + mj;
+      }
     }
     else
     {
-      ej = 1.0;
+      for (arma::uword i = 0; i < nrX; ++i)
+      {
+        col[i] += mj;
+      }
     }
-    // pass 1: unstandardize, accumulate weighted mean
-    double new_mean = 0.0;
-    double sum_sq = 0.0;
-    for (arma::uword i = 0; i < nrX; ++i)
+    // gather: accumulate only over missing rows
+    const arma::uword beg = miss_rows_offsets[cidx];
+    const arma::uword end = miss_rows_offsets[cidx + 1];
+    double imp_sum = 0.0;
+    double imp_sumsq = 0.0;
+    for (arma::uword p = beg; p < end; ++p)
     {
-      double val;
-      if constexpr (Scale)
-      {
-        val = col[i] * ej + mj;
-      }
-      else
-      {
-        val = col[i] + mj;
-      }
-      col[i] = val;
+      const arma::uword i = miss_rows_flat[p];
+      const double v = col[i];
       const double w = wptr[i];
-      new_mean += w * val;
+      imp_sum += w * v;
       if constexpr (Scale)
       {
-        sum_sq += w * val * val;
+        imp_sumsq += w * v * v;
       }
     }
-    // pass 2: re-center (+re-scale)
+    const double new_mean = obs_sum[j] + imp_sum;
+    // pass 2: re-standardize
     if constexpr (Scale)
     {
-      const double var = sum_sq - new_mean * new_mean;
-      const double new_et = std::sqrt(std::max(0.0, var));
+      const double new_var = obs_sumsq[j] + imp_sumsq - new_mean * new_mean;
+      const double new_et = std::sqrt(std::max(0.0, new_var));
       const double inv_et = 1.0 / std::max(new_et, PCA_TOL);
       for (arma::uword i = 0; i < nrX; ++i)
       {
@@ -243,9 +71,7 @@ void restandardize(arma::mat &Xhat,
     else
     {
       for (arma::uword i = 0; i < nrX; ++i)
-      {
         col[i] -= new_mean;
-      }
     }
     mean_p(j) = new_mean;
   }
@@ -328,8 +154,12 @@ Rcpp::List pca_imp_internal_cpp(
     const arma::rowvec &row_w,
     const double coeff_ridge)
 {
+  LOC_TIMER_OBJ(pca_imp_gram);
+  LOC_TIC(pca_imp_gram, "pca_imp_internal_cpp_total");
+
   const arma::uword nrX = X.n_rows;
   const arma::uword ncX = X.n_cols;
+  const arma::uword ntot = nrX * ncX;
   const double nrX_d = static_cast<double>(nrX);
   const double ncX_d = static_cast<double>(ncX);
   double old = arma::datum::inf;
@@ -374,12 +204,14 @@ Rcpp::List pca_imp_internal_cpp(
   arma::mat w_mask = not_miss;
   w_mask.each_col() %= row_w_col;
   const double *wmp = w_mask.memptr();
+  const double *xhp = Xhat.memptr();
+  const double *fxp = fittedX.memptr();
   // constants for SVD path
   const bool tall = (nrX >= ncX);
   const arma::uword k = ncp + 1; // clamped in R
   // loop-carried buffers. Note: no need to size upfront since armadillo sizes
   // them during the first iteration and subsequence iterations are no op
-  arma::mat U, V, X_work, AA_NxN, Vn, eigvecs, diff;
+  arma::mat U, V, X_work, AA_NxN, Vn, eigvecs;
   arma::vec vs_top, d_inv, eigvals, lambda_shrinked;
   // size of AA_NxN passed to the eig solver
   GramWorkspace gram_ws;
@@ -388,33 +220,106 @@ Rcpp::List pca_imp_internal_cpp(
     Rcpp::stop("workspace query failed");
   }
   AA_NxN.set_size(gram_ws.n, gram_ws.n);
+  // set sizes of other buffers
+  U.set_size(nrX, ncp);
+  V.set_size(ncX, ncp);
+  Vn.set_size(tall ? ncX : nrX, ncp);
+  fittedX.set_size(nrX, ncX);
+  if (tall)
+  {
+    X_work.set_size(nrX, ncX);
+  }
+  eigvecs.set_size(tall ? ncX : nrX, k);
+  eigvals.set_size(k);
+  vs_top.set_size(k);
+  d_inv.set_size(ncp);
+  lambda_shrinked.set_size(ncp);
   // pre-compute constants for sigma2
   const double ncp_d = static_cast<double>(ncp);
   const double min_dim = std::min(ncX_d, nrX_d - 1.0);
   const double denom_sigma = (ncX_d - ncp_d) * (nrX_d - 1.0 - ncp_d);
   const double scale_factor = (nrX_d * ncX_d) / min_dim;
-  // Rcpp::Timer timer("pca_imp_gram");
-  // Rcpp::Timer::ScopedTimer scpdtmr(timer, "pca_imp_internal_cpp_total"); // measure total time
+  // pre-compute per-column missing structure and observed weighted sums. This
+  // doesn't change so we can just calculate it once. Columns with zero missing
+  // entries are skipped entirely in restandardize
+  const arma::urowvec col_nmiss_row = arma::sum(miss, 0);
+  arma::uvec miss_cols_idx;
+  {
+    std::vector<arma::uword> tmp;
+    tmp.reserve(ncX);
+    for (arma::uword j = 0; j < ncX; ++j)
+    {
+      if (col_nmiss_row[j] > 0)
+      {
+        tmp.push_back(j);
+      }
+    }
+    miss_cols_idx = arma::uvec(tmp);
+  }
+  const arma::uword n_mc = miss_cols_idx.n_elem;
+  arma::uvec miss_rows_offsets(n_mc + 1);
+  miss_rows_offsets[0] = 0;
+  for (arma::uword c = 0; c < n_mc; ++c) {
+    miss_rows_offsets[c + 1] = miss_rows_offsets[c] + col_nmiss_row[miss_cols_idx[c]];
+  }
+  arma::uvec miss_rows_flat(miss_rows_offsets[n_mc]);
+  arma::vec obs_sum_vec(ncX, arma::fill::zeros);
+  arma::vec obs_sumsq_vec(ncX, arma::fill::zeros);
+  {
+    const double *xp_init = X.memptr();
+    const arma::uword *mp_init = miss.memptr();
+    for (arma::uword cidx = 0; cidx < n_mc; ++cidx)
+    {
+      const arma::uword j = miss_cols_idx[cidx];
+      const arma::uword off = j * nrX;
+      arma::uword p = miss_rows_offsets[cidx];
+      double os = 0.0, oss = 0.0;
+      for (arma::uword i = 0; i < nrX; ++i)
+      {
+        if (mp_init[off + i])
+        {
+          miss_rows_flat[p++] = i;
+        }
+        else
+        {
+          const double w = wptr[i];
+          const double x = xp_init[off + i];
+          os += w * x;
+          oss += w * x * x;
+        }
+      }
+      obs_sum_vec[j] = os;
+      obs_sumsq_vec[j] = oss;
+    }
+  }
+  const double *obs_sum_ptr = obs_sum_vec.memptr();
+  const double *obs_sumsq_ptr = obs_sumsq_vec.memptr();
+  // hot while loop
   for (arma::uword nb_iter = 1;; ++nb_iter)
   {
+    LOC_TIC(pca_imp_gram, "restandardize");
     Xhat.elem(missing) = fittedX.elem(missing);
-
-    // timer.tic("restandardize");
     if (scale)
     {
-      restandardize<true>(Xhat, mean_p, et, wptr, nrX, ncX);
+      restandardize<true>(Xhat, mean_p, et, wptr, nrX,
+                          miss_cols_idx, miss_rows_flat, miss_rows_offsets,
+                          obs_sum_ptr, obs_sumsq_ptr);
     }
     else
     {
-      restandardize<false>(Xhat, mean_p, et, wptr, nrX, ncX);
+      restandardize<false>(Xhat, mean_p, et, wptr, nrX,
+                           miss_cols_idx, miss_rows_flat, miss_rows_offsets,
+                           obs_sum_ptr, obs_sumsq_ptr);
     }
-    // timer.toc("restandardize");
-    // timer.tic("svd");
+    LOC_TOC(pca_imp_gram, "restandardize");
+
+    LOC_TIC(pca_imp_gram, "svd");
     SVD_triplet(Xhat, sptr, isptr, ncp, tall,
                 vs_top, trace_val, U, V, X_work, AA_NxN,
-                d_inv, Vn, eigvals, eigvecs, gram_ws);
-    // timer.toc("svd");
-    // timer.tic("post_svd");
+                d_inv, Vn, eigvals, eigvecs, gram_ws LOC_TIMER_ARG(pca_imp_gram));
+    LOC_TOC(pca_imp_gram, "svd");
+
+    LOC_TIC(pca_imp_gram, "post_svd");
     sigma2 = 0.0;
     if (regularized)
     {
@@ -426,26 +331,22 @@ Rcpp::List pca_imp_internal_cpp(
     const auto vs_h = vs_top.head(ncp);
     lambda_shrinked = (vs_h % vs_h - sigma2) /
                       arma::clamp(vs_h, PCA_TOL, arma::datum::inf);
-    // timer.toc("post_svd");
-    // timer.tic("reconstruct");
+    LOC_TOC(pca_imp_gram, "post_svd");
+
+    LOC_TIC(pca_imp_gram, "reconstruct+objective");
     scale_cols_inplace(U.memptr(), lambda_shrinked.memptr(), U.n_rows, ncp);
     fittedX = U * V.t();
-    // timer.toc("reconstruct");
-
-    // timer.tic("objective");
     {
-      const double *xp = Xhat.memptr();
-      const double *fp = fittedX.memptr();
-      const arma::uword ntot = nrX * ncX;
       double acc = 0.0;
       for (arma::uword idx = 0; idx < ntot; ++idx)
       {
-        const double d = xp[idx] - fp[idx];
+        const double d = xhp[idx] - fxp[idx];
         acc += wmp[idx] * d * d;
       }
       objective = acc;
     }
-    // timer.toc("objective");
+    LOC_TOC(pca_imp_gram, "reconstruct+objective");
+
     // convergence
     const double criterion = (old > 0.0 && !std::isinf(old))
                                  ? std::abs(1.0 - objective / old)
@@ -462,7 +363,7 @@ Rcpp::List pca_imp_internal_cpp(
       break;
     }
   }
-  // timer.tic("postprocessing");
+  LOC_TIC(pca_imp_gram, "postprocessing");
   // final unstandardize
   arma::vec imputed_vals(missing.n_elem);
   double sse = 0.0;
@@ -475,7 +376,9 @@ Rcpp::List pca_imp_internal_cpp(
     finalize_impute<false>(X, Xhat, fittedX, miss, et, mean_p, imputed_vals, sse);
   }
   const double mse = sse / static_cast<double>(X.n_elem - missing.n_elem);
-  // timer.toc("postprocessing");
+  LOC_TOC(pca_imp_gram, "postprocessing");
+
+  LOC_TOC(pca_imp_gram, "pca_imp_internal_cpp_total");
 
   return Rcpp::List::create(
       Rcpp::Named("imputed_vals") = imputed_vals,
