+++ src/mat_stats.cpp
@@ -1,4 +1,3 @@
-// [[Rcpp::depends(RcppArmadillo)]]
 #include <RcppArmadillo.h>
 #if defined(_OPENMP)
 #include <omp.h>
